Re-enable experimental support for cvmfs on macos runners
While macfuse is no longer an option, there is now a nice alternative with fuse-t.
This PR adds experimental support (I've successfully tested it with the nightly build, but there is no released cvmfs version
with fuse-t support yet).

'Synthetic firmlinks' are used to create /cvmfs , since the root file system is now read-only on macos.

One todo: need to check if we actually need the homebrew action or it can be moved somewhere else.